+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 [Unreleased]## [0.5.5.2b6]
+### Added
+- DeployedDOModel.monitor_execution efficiency improvement: only polls for the 'status'
 
 ## [0.5.5.2b5] - 2024-05-07
 ### Added